fix(floating-big-letter): replace whole buffer
There is always at least one, empty line in a buffer.
When replacing the range [0,0[, the new contents are inserted before
said empty line.
As of nvim 0.10, this results in the new contents being scrolled up by
one row, breaking the vertical centering.
Fix this by ensuring that the entire buffer content is overridden.